third_party/WebKit/LayoutTests/media/video-controls-always-visible-when-control-hovered.html

<!DOCTYPE html>
<title>Verify that media controls are always visible when hovered.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-file.js"></script>
<script src="media-controls.js"></script>
<video controls loop></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");

    video.oncanplaythrough = t.step_func(function() {

        assert_true(video.paused);

        // Click the play button.
        var playCoords = mediaControlsButtonCoordinates(video, "play-button");
        eventSender.mouseMoveTo(playCoords[0], playCoords[1]);
        eventSender.mouseDown();
        eventSender.mouseUp();
        assert_false(video.paused);

        runAfterHideMediaControlsTimerFired(t.step_func_done(function() {
            var controls = mediaControlsButton(video, "panel");
            assert_equals(getComputedStyle(controls).opacity, "1");
        }), video);
    });

    video.src = findMediaFile("video", "content/test");
});
</script>
